+++ src/Saloon/Data/Info/GetTimezoneResponseData.php
@@ -6,4 +6,20 @@ namespace GracjanKubicki\LaravelZadarma\Saloon\Data\Info;
 
 use GracjanKubicki\LaravelZadarma\Saloon\Data\ZadarmaResponseData;
 
-final class GetTimezoneResponseData extends ZadarmaResponseData {}
+final class GetTimezoneResponseData extends ZadarmaResponseData
+{
+    public function timezone(): ?string
+    {
+        return $this->string('timezone');
+    }
+
+    public function dateTime(): ?string
+    {
+        return $this->string('datetime');
+    }
+
+    public function unixTime(): ?int
+    {
+        return $this->integer('unixtime');
+    }
+}