+++ src/Saloon/Data/Info/ListTariffsResponseData.php
@@ -6,4 +6,13 @@ namespace GracjanKubicki\LaravelZadarma\Saloon\Data\Info;
 
 use GracjanKubicki\LaravelZadarma\Saloon\Data\ZadarmaResponseData;
 
-final class ListTariffsResponseData extends ZadarmaResponseData {}
+final class ListTariffsResponseData extends ZadarmaResponseData
+{
+    /**
+     * @return list<mixed>
+     */
+    public function tariffs(): array
+    {
+        return $this->listValue('tariffs');
+    }
+}