+++ src/Saloon/Data/Info/RequestCallbackResponseData.php
@@ -6,4 +6,10 @@ namespace GracjanKubicki\LaravelZadarma\Saloon\Data\Info;
 
 use GracjanKubicki\LaravelZadarma\Saloon\Data\ZadarmaResponseData;
 
-final class RequestCallbackResponseData extends ZadarmaResponseData {}
+final class RequestCallbackResponseData extends ZadarmaResponseData
+{
+    public function callId(): ?string
+    {
+        return $this->string('call_id');
+    }
+}